images: refactor - rename ImagePane to FeatureImages
