+++ src/components/FeaturePanel/FeatureImages/Image/PanoramaImg.tsx
