+++ src/components/FeaturePanel/FeatureImages/PathsSvg.tsx
