+++ src/components/FeaturePanel/FeatureImages/helpers.tsx
